+++ packages/components/molecules/f-media-element/src/components/_tests/MediaElement.test.js
@@ -56,75 +56,6 @@ describe('MediaElement.vue', () => {
             expect(textWrapper.text()).toEqual(mockText);
         });
 
-        it('should apply stack class when stacked prop is true', () => {
-            // Arrange
-            const wrapper = shallowMount(MediaElement, {
-                propsData: {
-                    title: mockTitle,
-                    text: mockText,
-                    imageUrl: mockImageUrl,
-                    stacked: true
-                },
-                mocks: {
-                    $style: {
-                        'c-mediaElement--stack': 'c-mediaElement--stack'
-                    }
-                }
-            });
-
-            // Act
-            const style = wrapper.find('.c-mediaElement--stack');
-
-            // Assert
-            expect(style.exists()).toBe(true);
-        });
-
-        it('should apply reverse class when reverse prop is true', () => {
-            // Arrange
-            const wrapper = shallowMount(MediaElement, {
-                propsData: {
-                    title: mockTitle,
-                    text: mockText,
-                    imageUrl: mockImageUrl,
-                    reverse: true
-                },
-                mocks: {
-                    $style: {
-                        'c-mediaElement--reverse': 'c-mediaElement--reverse'
-                    }
-                }
-            });
-
-            // Act
-            const style = wrapper.find('.c-mediaElement--reverse');
-
-            // Assert
-            expect(style.exists()).toBe(true);
-        });
-
-        it('should apply stackWhenNarrow class when stackWhenNarrow prop is true', () => {
-            // Arrange
-            const wrapper = shallowMount(MediaElement, {
-                propsData: {
-                    title: mockTitle,
-                    text: mockText,
-                    imageUrl: mockImageUrl,
-                    stackWhenNarrow: true
-                },
-                mocks: {
-                    $style: {
-                        'c-mediaElement--stackWhenNarrow': 'c-mediaElement--stackWhenNarrow'
-                    }
-                }
-            });
-
-            // Act
-            const style = wrapper.find('.c-mediaElement--stackWhenNarrow');
-
-            // Assert
-            expect(style.exists()).toBe(true);
-        });
-
         it.each([
             [ALIGN.LEFT, 'c-mediaElement-content--left'],
             [ALIGN.RIGHT, 'c-mediaElement-content--right'],
@@ -182,12 +113,12 @@ describe('MediaElement.vue', () => {
         });
 
         it.each([
-            [FONT_SIZE.SM, 'c-mediaElement-content--fontSizeSmall'],
-            [FONT_SIZE.MD, 'c-mediaElement-content--fontSizeMedium'],
-            [FONT_SIZE.LG, 'c-mediaElement-content--fontSizeLarge'],
-            [FONT_SIZE.XL, 'c-mediaElement-content--fontSizeXLarge'],
-            [FONT_SIZE.XXL, 'c-mediaElement-content--fontSizeXXLarge'],
-            ['__TEST_DUMMY_VALUE__', 'c-mediaElement-content--fontSizeMedium']
+            [FONT_SIZE.SM, 'c-mediaElement-contentFontSize--sm'],
+            [FONT_SIZE.MD, 'c-mediaElement-contentFontSize--md'],
+            [FONT_SIZE.LG, 'c-mediaElement-contentFontSize--lg'],
+            [FONT_SIZE.XL, 'c-mediaElement-contentFontSize--xl'],
+            [FONT_SIZE.XXL, 'c-mediaElement-contentFontSize--xxl'],
+            ['__TEST_DUMMY_VALUE__', 'c-mediaElement-contentFontSize--md']
         ])('should when textSize prop is %s, set class %s', (textSizeProp, textSizeClass) => {
             // Arrange
             const wrapper = shallowMount(MediaElement, {